Extracts I/O functionality to rake tasks; adds OpenSans fonts
This addresses two issues:

 1) We really need the import and export processes to be marshalled a
    little more heavily and for the import/ and export/ folders to be
more robust in their usage in this process.

 2) The current fonts are constrained to basic latin alphabets, and
    break with any UTF8 characters - unfortunately, even in English this
breaks easily when any non-standard quotation marks or other glyphs are
used (common in word processors)

The freely available and openly licensed OpenSans font is both
reasonably aesthetic and readily available. This has been added, and the
PDF Template has had this added as the default. There are three
variants, and each have subtypes defined for various styles. It will
"just work".

There are default rake tasks we get for free: :clobber and :clean, and
we can tap into them to help keep the import/ and export/ files tidy.
This has been configured in the Rakefile.

Separate tasks for import and export are defined in individual rake
extensions under lib/tasks. The import task pulls from a CSV source
file, converts it, and dumps it to a YML file in import/. The export
task, for now, just gathers all files in import/ and converts each of
them and exports the rendered PDF into the export/ folder.

Some additional formatting is done in the PDF to make more text fit onto
the cards and make them look a little nicer.